// src/app/common/error-message-generator/error-message-type.enum.ts
export enum ErrorMessageTypeEnum {
  max = 'max',
  min = 'min',
  maxLength = 'maxLength',
  minLength = 'minLength',
  minDate = 'minDate',
  maxDate = 'maxDate',
  required = 'required',
  uniqueName = 'uniqueName',
  integer = 'integer',
  decimal = 'decimal',
  pattern = 'pattern',
  email = 'email',
  time = 'time',
  date = 'date',
  dateTime = 'dateTime',
  maxCount = 'maxCount',
  minCount = 'minCount',
  invalidNumber = 'invalidNumber',
  whiteSpace = 'whiteSpace',
  numberMaxLength = 'numberMaxLength',
  decimalLimitExceeded = 'decimalLimitExceeded',
  invalidYoutubeUrl = 'invalidYoutubeUrl'
}